refactor(editor): consolidate editor files into src/ui/editor/ subdirectory
