+++ src/ui/RequestPane.tsx
@@ -10,22 +10,22 @@ import { useCallback, useEffect, useMemo, useRef, useState } from "react"
 import type { Auth, Request, Environment } from "../schema"
 import { formatBody } from "./formatRequest"
 import type { EditState, FieldKind } from "./editMode"
-import type { CodeEditorRenderable } from "./CodeEditor"
-import { CodeEditorCompletion } from "./CodeEditorCompletion"
+import type { CodeEditorRenderable } from "./editor/CodeEditor"
+import { CodeEditorCompletion } from "./editor/CodeEditorCompletion"
 
 import { CenterText } from "./CenterText"
 import { Tabs, type TabDef } from "./Tabs"
 import { useTheme } from "./theme"
 import type { Theme } from "./theme"
 import { FullBorder, LeftBar } from "./borders"
-import { JsonBodyViewer } from "./JsonBodyViewer"
+import { JsonBodyViewer } from "./editor/JsonBodyViewer"
 import { VarInput } from "./VarInput"
 import { KeyValueSection } from "./KeyValueSection"
 import { Checkbox } from "./Checkbox"
 import { AuthEditor } from "./AuthEditor"
 import { Select, type SelectItem } from "./Select"
 import { FormEditor } from "./FormEditor"
-import { ValidationNotice } from "./ValidationNotice"
+import { ValidationNotice } from "./editor/ValidationNotice"
 import type { BodyType } from "../schema"
 import { validateJsonContent } from "./jsonValidation"
 import { getEnvVarHighlights } from "./variableCompletion"
