+++ src/ui/editor/useJsonHighlight.ts
@@ -1,12 +1,12 @@
 import { SyntaxStyle } from "@opentui/core"
 import type { TextareaRenderable } from "@opentui/core"
-import type { Theme } from "../ui/theme-data"
-import { highlightJsonTokens } from "../ui/syntax"
-import type { Environment } from "../schema"
+import type { Theme } from "../theme-data"
+import { highlightJsonTokens } from "./syntax"
+import type { Environment } from "../../schema"
 import {
   buildCharToDisplayOffsets,
   charOffsetToDisplayOffset,
-} from "../ui/highlightOffsets"
+} from "../highlightOffsets"
 
 function createJsonSyntaxStyle(theme: Theme): SyntaxStyle {
   return SyntaxStyle.fromStyles({
