+++ src/ui/jsx-types.d.ts
@@ -1,4 +1,4 @@
-import type { CodeEditorRenderable } from "./CodeEditor"
+import type { CodeEditorRenderable } from "./editor/CodeEditor"
 
 declare module "@opentui/react" {
   interface OpenTUIComponents {
