+++ tests/unit/CodeEditorCompletion.test.tsx
@@ -6,8 +6,8 @@ import { testRender } from "@opentui/react/test-utils"
 import { createTestKeymap } from "@opentui/keymap/testing"
 import { KeymapProvider } from "@opentui/keymap/react"
 import type { KeymapProviderProps } from "@opentui/keymap/react"
-import { CodeEditorRenderable } from "../../src/ui/CodeEditor"
-import { CodeEditorCompletion } from "../../src/ui/CodeEditorCompletion"
+import { CodeEditorRenderable } from "../../src/ui/editor/CodeEditor"
+import { CodeEditorCompletion } from "../../src/ui/editor/CodeEditorCompletion"
 import { VariableCompletionInterceptor } from "../../src/ui/variableCompletionInterceptor"
 import { ThemeProvider } from "../../src/ui/theme"
 import type { Environment } from "../../src/schema"
