+++ tests/unit/YamlEditorOverlay.test.tsx
@@ -9,9 +9,9 @@ import {
 import { KeymapProvider } from "@opentui/keymap/react"
 import type { KeymapProviderProps } from "@opentui/keymap/react"
 import { ThemeProvider } from "../../src/ui/theme"
-import { YamlEditorOverlay } from "../../src/ui/YamlEditorOverlay"
+import { YamlEditorOverlay } from "../../src/ui/editor/YamlEditorOverlay"
 import { ConfirmOverlay } from "../../src/ui/ConfirmOverlay"
-import { CodeEditorRenderable } from "../../src/ui/CodeEditor"
+import { CodeEditorRenderable } from "../../src/ui/editor/CodeEditor"
 
 extend({ "code-editor": CodeEditorRenderable })
